# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Student Hobby Survey
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SurveySubmission {
  major: string;
  state: string;
  frequency: string;
  hobbies: string[];
  other_hobby?: string;
}

export interface SurveyResponse {
  id: string;
  major: string;
  state: string;
  frequency: string;
  hobbies: string[];
  other_hobby?: string;
  created_at: string;
}

export interface FrequencyCount {
  frequency: string;
  count: number;
}

export interface HobbyCount {
  hobby: string;
  count: number;
}

export interface StateCount {
  state: string;
  count: number;
}

export interface SurveyResults {
  total_responses: number;
  frequency_counts: FrequencyCount[];
  hobby_counts: HobbyCount[];
  top_states: StateCount[];
}

export interface ErrorResponse {
  error: string;
}