# lib/api-zod/src/generated/types/hobbyCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Student Hobby Survey
 * OpenAPI spec version: 0.1.0
 */

export interface HobbyCount {
  hobby: string;
  count: number;
}