# lib/api-zod/src/generated/types/surveyResults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Student Hobby Survey
 * OpenAPI spec version: 0.1.0
 */
import type { FrequencyCount } from "./frequencyCount";
import type { HobbyCount } from "./hobbyCount";
import type { StateCount } from "./stateCount";

export interface SurveyResults {
  total_responses: number;
  frequency_counts: FrequencyCount[];
  hobby_counts: HobbyCount[];
  top_states: StateCount[];
}