# lib/api-zod/src/generated/types/surveySubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Student Hobby Survey
 * OpenAPI spec version: 0.1.0
 */

export interface SurveySubmission {
  major: string;
  state: string;
  frequency: string;
  hobbies: string[];
  other_hobby?: string;
}